src/requirements.txt
certifi>=2025.1.31
charset-normalizer>=3.4.1
docopt>=0.6.2
idna>=3.10
pillow>=11.1.0
PyQt6>=6.8.1
PyQt6-Qt6>=6.8.2
PyQt6_sip>=13.10.0
requests>=2.32.3
urllib3>=2.3.0
yarg>=0.1.10
